Update to edition 2018 and compile with recent rust versions
This commit contains a lot of small changes to update fall to edition
2018, mostly adding lots of dyn to dynamic object references and
updating imports.

Tera is updated to a recent version since older version of tera used a
version of the url crate that did not compile anymore.

An update to the fst crate made the fst::Map generic. The previous
implicit Vec<u8> is used as the argument now.

The generated tests are updated to also be valid rust 2018.

Test output ocmparisons are now done trimming the ends of lines, in
order to work when using a CLion that insists on stripping trailing
whitespace.